refactor(core): extract windowed word decode from Engine
Move decode_words, decode_speech_regions, decode_words_streaming,
and decode_words_for_samples into engine/words.rs. Public
transcribe_* wrappers stay in transcribe.rs.
