+++ CHANGELOG.md
@@ -25,7 +25,7 @@ were released without a git tag, so their headings carry no compare link.
 
 - Windows `cargo test --workspace --lib --bins` CI job.
 - Nightly AddressSanitizer job over the session pool (alongside Miri + TSAN).
-- `Engine` impl split into `engine/{config,load,stream,transcribe,infer}.rs`.
+- `Engine` impl split into `engine/{config,load,stream,transcribe,infer,words}.rs`.
   Public API is unchanged.
 - `model` download/identity split into `model/{progress,variant,download}.rs`;
   the streaming SHA-256 fetch lives in `download/fetch.rs`, ANE
