+++ crates/gigastt-core/src/inference/engine/mod.rs
@@ -178,3 +178,4 @@ mod stream;
 #[cfg(test)]
 mod tests;
 mod transcribe;
+mod words;
